Bluetooth-Wipower: Fix for BLE GATT Server disconnection failure.
Create dummy connection to hold the connection even after stop
advertisments for wipower is restricted to wipower profile
only, Earlier it was creating a dummy connection during for all
the new gatt connections.
Fix for low RSSI, Wipower needs the tx power level to be high
so that there is not drop in RSSI level when on PTU, As part
"set multiple advertising parameters" the paramters sent earlier
was  with low power which reduced 21DB in RSSI so now its been
changed for default RSSI levels.

Change-Id: Ib6015d2c9a23c56358ea69649dd475675377ef5b
CRs-fixed: 781789
CRs-fixed: 781824